Load Designer JS data update
In javascript section of main file that load Designer UI
for 'var credentials' now we're sending the access token and refresh token, instead access code, and client credentials.

This is just because currently a lot of access tokens were being created for each process design is open, and while it is being open is sending two request, the first one to exchange an access code by an access token,..

Now, whith this change, while a process design is open we will made just one request to some resource with access token in the header,.. exchange access code by an access token is not needed any more.

NOTE.- the access token passed to the front-end can be expired at first time,.. so we need that the exchange an access token by a refresh token handling is working well.