BUG-14953 WEBENTRY - ProcessMaker does not check if Javascript is running.
When JS is not enabled in the browser, the JS validations dosen`t work, so you can submit the form even if there are empty requiered fields.

A validation <noscript> was added in the HTML code, to validate if the browser has or not JS enable.

If there is not JS enable, the validation is make in the php code to.