BUG-15608 Suggest field doesn't display the suggestion list in the third letter.
A suggest field with the type of search: "Result that contain the entered text", in the third letter that is enter the suggestion list is not displayed.

The problem whit the suggestion field was in a js validation, when the function indexOf() was used to return the position of the string that was search, this was happening because the validation was comparing with '0' when the string was not found, when this function returns -1 and not 0 if the item is not found, 0 was the initial position in the string.